Add example of using synchronized
synchronized - protection of code from simultaneous using by more than one thread.

when one thread goes in synchronized method or block it takes monitor of the object. Other threads cannot wait until the monitor is released.

It can be said that synchronized made method atomic.

All modifications made by thread which took the monitor are guaranteed to be visible to other objects.

In comparison with volatile synchronized is heavier way of protection, because operations of taking and releasing of a monitor reduce performance.

A synchronized block is sometimes better because it allows you to:

1.Synchronize on a different object
2.Limit the scope of synchronization

So if you want to lock the whole object, use can use eiter a synchronized method or block. If you want to keep other parts of the object accessible to other threads, use synchronized block